+++ changelog.md
@@ -8,7 +8,7 @@
 
 ### Internals
 
-- refactored timeout handling in RequestResponse and ValueReader.
+- refactored timeout handling in GatewayScanner, RequestResponse and ValueReader.
 - renamed "PhysicalAddress" to "IndividualAddress"
 - Telegram: `group_address` renamed to `destination_address`, to prepare support for other APCI services and add `source_address`
 - Telegram: remove `Telegram.telegramtype` and replace with payload object derived from `xknx.telegram.apci.APCI`.